refactor: :recycle: ディレクトリ構成変更
