+++ src/app/Http/Controllers/Admin/AdminController.php
@@ -1,11 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
-use Illuminate\Http\Request;
-
-use Illuminate\Support\Facades\Auth;
-
+use App\Http\Controllers\Controller;
 use App\Models\BigQuestion;
 
 class AdminController extends Controller
